// frontend/src/core/enums/errors.enum.ts
export enum Errors {
  BadRequest = 'Bad Request',
  Unauthorized = 'Unauthorized',
  Forbidden = 'Forbidden',
  NotFound = 'Not Found',
  MethodNotAllowed = 'Method Not Allowed',
  NotAcceptable = 'Not Acceptable',
  ProxyAuthenticationRequired = 'Proxy Authentication Required',
  RequestTimeout = 'Request Timeout',
  Conflict = 'Conflict',
  Gone = 'Gone',
  LengthRequired = 'Length Required',
  UnprocessableEntity = 'Unprocessable Entity',
  TooManyRequests = 'Too Many Requests',
  InternalServerError = 'Internal Server Error'
}